Switch from rbtree to rbtree3
The original rbtree library has not been updated since 2013.  It has
compatibility problems with recent versions of Ruby.  It produces
deprecation warnings on Ruby 2.7 and is not friendly with MJIT and
Compaction GC, and it does not even build with Ruby 3.0 soon to be
released.

To fix them, I suggest switching to the fork called
[rbtree3](https://github.com/kyrylo/rbtree3).  It is used by the [ruby
client](https://github.com/airbrake/airbrake-ruby) of
[Airbrake](https://airbrake.io/) and developed by the vendor.

I confirmed it builds and works with the latest snapshot of Ruby 3.0.0dev.